fix(fex): resolve libFEXCore.dylib at build time, fix Windows format-string errors
Two independent CI failures on the first push:

- Windows clang-tidy (-Werror,-Wformat): audio_service.cpp and
  service_control.cpp logged several ULONG values with "%u". ULONG is
  a distinct type from unsigned int under Windows' LLP64 model (even
  though same width), so clang flags it. Cast each value to uint32_t
  at the call site to match the format string.
- macOS arm64 Python wheel build: the Python extension module needs to
  dlopen the freshly-built module at build time (to generate .pyi
  stubs), which failed with "Library not loaded: @rpath/libFEXCore.dylib".
  FEXCore's own build gives its shared library the install name
  "@rpath/libFEXCore.dylib", but nothing pointed an rpath at the
  ExternalProject's actual build directory - every local build only
  "worked" because of a manual `cp .../libFEXCore.dylib
  build/release/artifacts/` step (still documented, until now, in
  docs/fex-backend.md) that a fresh/CI build never performs, and that
  a Python packaging build (a different build-tree layout entirely)
  has no equivalent of. Fixed by setting fex-emulator's BUILD_RPATH/
  INSTALL_RPATH directly to the fexcore imported target's own real
  location, derived at configure time - works unconditionally,
  independent of which build directory layout the consumer uses.
  Verified locally by removing the manually-copied dylib entirely and
  confirming the analyzer still resolves it via the new rpath.
  Updated docs/fex-backend.md's now-obsolete manual-copy note to
  match.

+++ src/windows-emulator/ports/audio_service.cpp
@@ -134,13 +134,14 @@ namespace sogen
                 if (getenv("EMULATOR_LOG_RPC"))
                 {
                     win_emu.log.error("[audiosrv] call iface=%02x%02x%02x%02x opnum=%u send=%u\n", iface[0], iface[1], iface[2], iface[3],
-                                      procedure_id, c.send_buffer_length);
+                                      procedure_id, static_cast<uint32_t>(c.send_buffer_length));
                 }
                 // DEBUG-AUDIO-NDR: dump full request bytes for the AudioClient streaming opnums under scrutiny.
                 if (getenv("EMULATOR_LOG_RPC") && this->is_audio_client_ &&
                     (procedure_id == 0 || procedure_id == 2 || procedure_id == 4 || procedure_id == 5 || procedure_id == 7))
                 {
-                    win_emu.log.error("[audiosrv-dbg] REQ opnum=%u send_len=%u bytes: %s\n", procedure_id, c.send_buffer_length,
+                    win_emu.log.error("[audiosrv-dbg] REQ opnum=%u send_len=%u bytes: %s\n", procedure_id,
+                                      static_cast<uint32_t>(c.send_buffer_length),
                                       dump_hex(win_emu, c.send_buffer, c.send_buffer_length, 256).c_str());
                 }
 
@@ -184,8 +185,9 @@ namespace sogen
 
             static NTSTATUS log_unhandled(windows_emulator& win_emu, const char* iface, const uint32_t opnum, const lpc_request_context& c)
             {
-                win_emu.log.error("[audiosrv] UNHANDLED %s opnum=%u send_len=%u recv_len=%u req: %s\n", iface, opnum, c.send_buffer_length,
-                                  c.recv_buffer_length, dump_hex(win_emu, c.send_buffer, c.send_buffer_length).c_str());
+                win_emu.log.error("[audiosrv] UNHANDLED %s opnum=%u send_len=%u recv_len=%u req: %s\n", iface, opnum,
+                                  static_cast<uint32_t>(c.send_buffer_length), static_cast<uint32_t>(c.recv_buffer_length),
+                                  dump_hex(win_emu, c.send_buffer, c.send_buffer_length).c_str());
                 return STATUS_NOT_SUPPORTED;
             }
